Fix cast ptr

+++ libft/srcs/ft_myputchar.c
@@ -14,5 +14,5 @@
 
 int			ft_myputchar(int c)
 {
-	return (write(1, &c, 1));
+	return (int) (write(1, &c, 1));
 }
